refactor: project card folder created
